Add RegionalSirModel.
This commit refactors the SimSirModel to allow overriding of the way
the model class calculates the infected population, and creates a
RegionalSirModel that uses a parameterized detection probabilty
instead of using the current_hospitalized. This is necessary for
region-wide calculations as the data for patients currently
hospitalized is not currently available.